\name{ADP-package}

\alias{ADP-package}

\alias{ADP}

\docType{package}

\title{
ADP: DDI Codebook From 1ka
}

\description{
Function to export data in Excel, SPSS and Stata.
}

\details{
    \tabular{ll}{
    Package: \tab ADP\cr
    Type:    \tab Package\cr
    Version: \tab 0.8\cr
    Date:    \tab 2024-08-21\cr
    License: \tab GPL (>= 2)\cr
    }
}

\author{

\bold{Authors}:\cr
Adrian Dusa\cr
\email{dusa.adrian@gmail.com}

\bold{Maintainer}:\cr
Adrian Dusa
}

\keyword{package}

