% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADP.R
\docType{data}
\name{ADP}
\alias{ADP}
\title{How many people use my product}
\format{
An object of class \code{data.frame} with 100 rows and 3 columns.
}
\source{
Kislev, Mickey M. & Kislev, Shira, (2020). The Market Trajectory of a
Radically New Product: E-Cigarettes. IJMS 12(4):63-92, DOI:\href{https://www.ccsenet.org/journal/index.php/ijms/article/view/0/44285}{10.5539/ijms.v12n4p63}
}
\usage{
ADP
}
\description{
A data base of adoption probability by triers and users
user_p - a vector of percentage  (0<users<1) of users.
triers_p - a vector of percentage  (0<triers<1) of triers
ADP - vector of predicted percentage (0<ADP<1) of the adoption probability
of an innovative product in the population.
}
\details{
The measuring of triers is relatively easy. It is just a question of whether
a person tried a product even once in his life or not. While measuring the
rate of people who also adopt it as part of their life is more complicated
since the adoption of a product is a subjective view of the individual.
Mickey Kislev and Shira Kislev developed a formula to calculates the prevalence
of users of a product to overcome this difficulty. The current dataseet
assists in calculating the users of a product based on the prevalence of
triers in the population.

For example, suppose that a candy company launched a new chocolate bar.
A candy company can collect data on the number of people who tried their
chocolate bar and know from the rate of triers how many people decided to
consume the new chocolate bar regularly. It should be noticed that the model
was proved only on consumer behaviour of adults above the age of 21 years old
and above.
}
\seealso{
\code{\link{ptriers}}, \code{\link{pusers}}, \code{\link{adp.t}}, and \code{\link{adp.u}}
}
\author{
Mickey Kislev and Shira Kislev
}
\keyword{datasets}
